Rearrange some things to better support thrift transport
